wait multiple times, add "closer heapq"
I want to be able to wait for data multiple times at mon_mitm prioQ
events to efficiently pick up all encounters for optimized IV scanning.
We need to move around a bit at the destination to keep the data coming
in 10 seconds steps instead of 30 seconds. That part needs to be made
prettier, it's not more than a little hack with forced values at this
point.

Also, workers can now stack a couple of events in their "closer_heapq"
for which they were found to be closest nearby. That way, they should - in
theory - jump lesser distances, remain within a certain "neighborhood"
instead of randomly jumping all over the area with every other event.
That should increase overall efficiency a bit.